requirements.txt
Django==1.7.7
unittest2==0.5.1
resolver==0.2.1
selector==0.8.11
httplib2==0.9
restclient==0.11.0
feedparser==5.1.3
Markdown==2.6.1
pyenchant==1.5.1
simplejson==3.6.5
smartypants==1.6.0.2
uuid==1.30
psycopg2==2.5.4
statsd==3.0.1
coverage==3.7.1
mccabe==0.2.1
pyflakes==0.8.1
pep8==1.5.7
flake8==2.2.5
logilab-common==0.59.0
logilab-astng==0.24.1
pylint==1.4.1
sqlparse==0.1.14
six==1.9.0
boto==2.38.0
python-dateutil==2.4.2
gunicorn==19.3.0
factory_boy==2.5.2

djangowind==0.13.2
template_utils==0.4p2
typogrify==1.0
django-tinymce==1.5.3
johnny-cache==1.6a
django-templatetag-sugar==0.1
django-paging==0.2.2
django-indexer==0.2.1
django-appconf==0.6
django-compressor==1.5
django-modeldict==1.1.6
django-jsonfield==0.5
django-statsd-mozilla==0.3.14
raven==5.2.0
django-annoying==0.8.0
django-smoketest==0.3.0
django-debug-toolbar==1.2.2
django-jenkins==0.17.0
django-waffle==0.10.1
django-ga-context==0.1.0
django-stagingcontext==0.1.0
django-impersonate==0.9
django-markwhat==1.3
django-storages-redux==1.2.2
django-cacheds3storage==0.1.1
